PagerSlidingTabStrip upgrade
If the adapter's dataset changes, this View does not reflect the changes
before onPageScrolled is called. This causes an NPE because
tabsContainer has no children. With this fix, the view is now synced
with the ViewPager's adapter (the correct tabs are present) and
tabsContainer#getChildAt will no longer throw an NPE.

Fixes #66

Conflicts:
	src/com/android/settings/rascarlo/PagerSlidingTabStrip.java